requirements.txt
# Core Framework
crewai>=0.30.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Trading APIs
oandapyV20>=0.6.3
requests>=2.31.0
websockets>=11.0.0

# Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.17.0
dash>=2.14.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Logging & Monitoring
structlog>=23.1.0
rich>=13.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0